# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CCIR API
  version: 1.0.0
  summary: Civic Complaint and Infrastructure Reporting backend.
  description: >
    Citizens file complaints about public infrastructure; agency staff and administrators triage,
    assign and resolve them.


    **Sessions.** Signing in sets two signed, HTTP-only cookies: `accessToken` (short-lived) and
    `refreshToken`. A request whose access token has expired is renewed from the refresh token
    automatically. There is no bearer-token mode.


    **Browser origin.** Every POST, PUT, PATCH and DELETE must carry an `Origin` header equal to the
    configured `BROWSER_ORIGIN`; otherwise the response is 403 `FORBIDDEN` ("Request origin is not
    allowed").


    **Errors.** Every error uses one envelope: `{ error: { code, message, details? }, msg }`.
    Unexpected server errors (500) also carry `error.requestId`, which matches the `X-Request-Id`
    response header and the server log line.


    **Request IDs.** Every response carries `X-Request-Id`. A caller may send its own
    `X-Request-Id` (8–64 characters of `A-Z a-z 0-9 . _ -`); anything else is replaced.


    **Rate limits.** 300 requests per 15 minutes for each signed-in account, and 300 for each
    client address without a valid session; all requests from one client address are also capped
    at 3000 per 15 minutes. Health probes are exempt. Sign-in and session refresh have tighter limits.
...
  license:
    name: MIT
    identifier: MIT
servers:
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: Signed HTTP-only access-token cookie set by sign-in; renewed automatically from the `refreshToken` cookie.
  headers:
    RequestId:
      description: Request identifier; quote it when reporting a problem.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{8,64}$'
...
    RetryAfter:
      description: Seconds until the limit resets.
      schema:
        type: integer
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: A 24-character hexadecimal identifier.
      schema:
        $ref: '#/components/schemas/ObjectId'
...
  schemas:
    ObjectId:
      type: string
      pattern: '^[0-9a-fA-F]{24}$'
...
    Error:
      type: object
      additionalProperties: false
      required: [error, msg]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [path, code, message]
                properties:
                  path:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
            requestId:
              type: string
              description: Present on 500 responses only.
        msg:
          type: string
    Message:
      type: object
      additionalProperties: false
      required: [msg]
      properties:
        msg:
          type: string
...
  responses:
    ValidationFailed:
      description: The request failed validation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details:
                - path: body.email
                  code: INVALID_VALUE
                  message: Invalid email address
            msg: Request validation failed
    Unauthorized:
      description: Not signed in, or the session is invalid or expired.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: Session expired, please log in again
            msg: Session expired, please log in again
    Forbidden:
      description: The signed-in user may not do this, or the request origin is not allowed.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action
            msg: You do not have permission to perform this action
    NotFound:
      description: No such resource.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: No complaint found with id 66f1a0c0a1b2c3d4e5f60001
            msg: No complaint found with id 66f1a0c0a1b2c3d4e5f60001
    Conflict:
      description: The request conflicts with current state (a duplicate, a stale version, or a rule such as a category still in use).
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: STALE_COMPLAINT
              message: This report changed; reload it and try again
            msg: This report changed; reload it and try again
...
    RateLimited:
      description: Too many requests.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RATE_LIMITED
              message: Too many requests, please try again later
            msg: Too many requests, please try again later
    ServiceUnavailable:
      description: A dependency needed for this request is unavailable or not configured.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: LOCATION_PROVIDER_UNAVAILABLE
              message: Location service is temporarily unavailable
            msg: Location service is temporarily unavailable
    ServerError:
      description: An unexpected error. The body never contains internal details; quote `requestId` when reporting it.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: Something went wrong
              requestId: 3f6c2a1e-8d4b-4c7a-9e2f-1b5d7c9a0e11
            msg: Something went wrong
...
security:
  - cookieAuth: []
paths:
...
  /auth/password/forgot:
    post:
      tags: [auth]
      summary: Ask for a password reset link
      description: >
        Always answers 202 with the same message, whether or not an account uses the address, and sends
        any email after answering, so neither the answer nor its timing reveals whether an account exists.
        An active password account gets a single-use link that expires in 30 minutes; a Google account is
        told to sign in with Google; suspended, retired and unknown addresses get nothing. Limited to 10
        requests per 15 minutes per client address and 3 per hour per email address.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 254
            example:
              email: ada@example.test
      responses:
        '202':
          description: Accepted.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                msg: If an account uses this address, we have sent it a link to reset the password.
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/password/reset:
    post:
      tags: [auth]
      summary: Set a new password with a reset link
      description: >
        The token comes from the link's address fragment (`#token=…`). The new password takes effect, every
        session of the account ends, and the account is emailed a notice; the caller is not signed in. An
        unknown, expired, used or replaced token gets 400 `INVALID_OR_EXPIRED_TOKEN`; a password equal to
        the email address gets 400 `PASSWORD_REJECTED` and leaves the link usable.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token, password]
              properties:
                token:
                  type: string
                  minLength: 1
                  maxLength: 100
                password:
                  type: string
                  minLength: 8
                  maxLength: 128
            example:
              token: 3q2-7wEXAMPLEexampleEXAMPLEexampleEXAMPLE0
              password: Brand-new-pass
      responses:
        '200':
          description: Password reset; sign in with the new password.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                msg: Password reset. Sign in with your new password.
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /auth/email/confirm:
    post:
      tags: [auth]
      summary: Confirm an email address change
      description: >
        The token comes from the link's address fragment (`#token=…`). Works signed in or out. The account's
        address changes, sessions are kept, and pending links for the account are cancelled. 400
        `INVALID_OR_EXPIRED_TOKEN`; 409 if the address was taken meanwhile or the account can no longer change.
      operationId: confirmEmailChange
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token:
                  type: string
                  minLength: 1
                  maxLength: 100
            example:
              token: 3q2-7wEXAMPLEexampleEXAMPLEexampleEXAMPLE0
      responses:
        '200':
          description: Email address changed.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                msg: Email address changed
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /users/profile/password:
    post:
      tags: [users]
      summary: Change the signed-in user's password
      description: >
        Needs the current password. A wrong one gets 401 `WRONG_PASSWORD` and counts against the account's
        sign-in lockout (5 per 15 minutes, then 429). The new password must be 8 to 128 characters, differ
        from the current one (400 `SAME_PASSWORD`) and not be the email address (400 `PASSWORD_REJECTED`).
        This session stays signed in; every other session ends, pending reset links are cancelled, and the
        account is emailed a notice. Google accounts have no password (409 `GOOGLE_ACCOUNT`).
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  minLength: 1
                  maxLength: 128
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 128
            example:
              currentPassword: Example-pass-1
              newPassword: Brand-new-pass
      responses:
        '200':
          description: Password changed.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                msg: Password changed
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/profile/email:
    post:
      tags: [users]
      summary: Ask to change the signed-in user's email address
      description: >
        Needs the current password (401 `WRONG_PASSWORD`, counted against the sign-in lockout). A single-use
        confirmation link, valid for 24 hours, goes to the new address and the old address is told; nothing
        changes until the link is used (`POST /auth/email/confirm`). 400 `SAME_EMAIL`, 409 when the address
        is in use or the account signs in with Google (`GOOGLE_ACCOUNT`), 503 `EMAIL_NOT_SENT` when the
        confirmation could not be sent (no link is kept; try again). 5 requests per hour per account.
      operationId: requestOwnEmailChange
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [newEmail, currentPassword]
              properties:
                newEmail:
                  type: string
                  format: email
                  maxLength: 254
                currentPassword:
                  type: string
                  minLength: 1
                  maxLength: 128
            example:
              newEmail: ada.new@example.test
              currentPassword: Example-pass-1
      responses:
        '202':
          description: Confirmation sent to the new address.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                msg: We have sent a confirmation link to the new address
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /users/{id}/email:
    post:
      tags: [users]
      summary: Ask to change a user's email address (administrators)
      description: >
        Sends a single-use confirmation link, valid for 24 hours, to the new address and tells the old one;
        nothing changes until the user confirms. Administrators change their own address from their profile
        (400 `USE_PROFILE`). 409 for a retired account, an address in use, or a Google account
        (`GOOGLE_ACCOUNT`); 503 `EMAIL_NOT_SENT` when the confirmation could not be sent.
      operationId: requestUserEmailChange
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [newEmail]
              properties:
                newEmail:
                  type: string
                  format: email
                  maxLength: 254
            example:
              newEmail: correct.address@example.test
      responses:
        '202':
          description: Confirmation sent to the new address.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                msg: We have sent a confirmation link to the new address
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...